rename useProposalsVotingResults->useProposalsVotingState
